[Refacter] dto 보기쉽게 몰아주기
