--- src/main/java/verbly/spring/domain/review/dto/request/QuizAnswerSubmitRequest.java
@@ -1,8 +1,0 @@
-package verbly.spring.domain.review.dto.request;
-
-import com.fasterxml.jackson.databind.JsonNode;
-
-public record QuizAnswerSubmitRequest(
-        JsonNode userAnswerJson,
-        String mistakeNote
-) {}
