--- src/main/java/verbly/spring/domain/review/dto/request/QuizHintRequest.java
@@ -1,3 +1,0 @@
-package verbly.spring.domain.review.dto.request;
-
-public record QuizHintRequest() {}
