+++ src/main/java/verbly/spring/global/common/code/SuccessStatus.java
@@ -27,7 +27,15 @@ public enum SuccessStatus implements BaseCode {
     CORRECTION_DELETE_SUCCESS(HttpStatus.OK, "CORRECTION2004", "Correction - 글을 성공적으로 삭제했습니다."),
     CORRECTION_BOOKMARK_ADD_SUCCESS(HttpStatus.OK, "CORRECTION2005", "Correction - 글을 성공적으로 즐겨찾기에 추가했습니다."),
     CORRECTION_BOOKMARK_REMOVE_SUCCESS(HttpStatus.OK, "CORRECTION2006", "Correction - 글을 즐겨찾기에서 성공적으로 삭제했습니다"),
-    ;
+    // 라이브러리 관련 응답
+    LIBRARY_ITEM_CREATE_SUCCESS(HttpStatus.CREATED, "LIB2001", "라이브러리 아이템이 생성되었습니다."),
+    LIBRARY_ITEM_LIST_SUCCESS(HttpStatus.OK, "LIB2002", "라이브러리 목록을 성공적으로 조회했습니다."),
+    LIBRARY_ITEM_DETAIL_SUCCESS(HttpStatus.OK, "LIB2003", "라이브러리 상세를 성공적으로 조회했습니다."),
+    LIBRARY_ITEM_UPDATE_SUCCESS(HttpStatus.OK, "LIB2004", "라이브러리 아이템이 수정되었습니다."),
+    LIBRARY_ITEM_DELETE_SUCCESS(HttpStatus.OK, "LIB2005", "라이브러리 아이템이 삭제되었습니다."),
+    LIBRARY_ITEM_EXAMPLE_ADD_SUCCESS(HttpStatus.CREATED, "LIB2006", "예문이 추가되었습니다."),
+    LIBRARY_ITEM_EXAMPLE_DELETE_SUCCESS(HttpStatus.OK, "LIB2007", "예문이 삭제되었습니다.");
+
 
     private final HttpStatus httpStatus;
     private final String code;
