+++ src/main/resources/application.yml
@@ -28,7 +28,7 @@ spring:
         format_sql: true
         use_sql_comments: true
         hbm2ddl:
-          auto: create #create #update #validate
+          auto: update #create #update #validate
         default_batch_fetch_size: 1000
   data:
     web:
